+++ common/src/test/scala/react/common/CommonTests.scala
@@ -1,5 +1,7 @@
 package react.common
 
+import react.common.syntax._
+import react.common.implicits._
 import cats.implicits._
 import scala.scalajs.js
 import utest._
@@ -26,6 +28,11 @@ object CommonTests extends TestSuite with TestUtils {
       val c: js.Object = js.Dynamic.literal(foo = 15, bar = "foobar")
       val m            = merge(a, b)
       assert(c === m)
+    }
+    'style - {
+      val a: js.Object = js.Dynamic.literal(height = 42, background = "black")
+      val b: js.Object = Style(Map("height" -> 42, "background" -> "black")).toJsObject
+      assert(a === b)
     }
   }
 }