+++ test/src/main/scala/react/common/TestUtils.scala
@@ -3,10 +3,6 @@ package react
 package common
 
 import japgolly.scalajs.react.ReactDOMServer
-import scala.scalajs.js
-import cats.Eq
-import cats.Show
-import cats.syntax.eq._
 import japgolly.scalajs.react.vdom.html_<^._
 import japgolly.scalajs.react.raw.React
 import japgolly.scalajs.react.vdom.{TagOf, TopNode}
@@ -14,45 +10,6 @@ import org.scalajs.dom.Element
 import utest._
 
 trait TestUtils { self: TestSuite =>
-  implicit def jsUndefOr[A: Eq]: Eq[js.UndefOr[A]] = Eq.instance { (a, b) =>
-    (a.toOption, b.toOption) match {
-      case (Some(a), Some(b)) => a === b
-      case _                  => false
-    }
-  }
-
-  implicit val eq: Eq[js.Object] = Eq.instance { (a, b) =>
-    val aDict = a.asInstanceOf[js.Dictionary[Any]]
-    val bDict = b.asInstanceOf[js.Dictionary[Any]]
-    (aDict.keySet == bDict.keySet) &&
-    aDict.keySet.forall(key => aDict(key) === bDict(key))
-  }
-
-  implicit val show: Show[js.Object] = Show.show { a =>
-    val aDict = a.asInstanceOf[js.Dictionary[Any]]
-    aDict.keySet.map(key => s"$key=${aDict(key)}").mkString("{", ",", "}")
-  }
-
-  implicit val jsEq: Eq[Any] = Eq.instance { (a, b) =>
-    (a, b) match {
-      case (a: js.Array[_], b: js.Array[_]) =>
-        a.length == b.length &&
-          a.zip(b).forall { x =>
-            jsEq.eqv(x._1, x._2)
-          }
-
-      case _
-          if a.asInstanceOf[js.Dynamic].constructor == js.constructorOf[js.Object] &&
-            b.asInstanceOf[js.Dynamic].constructor == js.constructorOf[js.Object] =>
-        val aDict = a.asInstanceOf[js.Dictionary[Any]]
-        val bDict = b.asInstanceOf[js.Dictionary[Any]]
-        (aDict.keySet == bDict.keySet) &&
-        aDict.keySet.forall(key => aDict(key) === bDict(key))
-
-      case _ =>
-        a == b
-    }
-  }
 
   def assertRender(e: VdomElement, expected: String): Unit =
     assertRender(e.rawElement, expected)