+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changes
 
+## 0.0.7
+
+* Add syntax for style
+* Rename the org name and split the packages
+
+## 0.0.6
+
+* Release with sbt-release-early
+
 ## 0.0.5
 
 * Add merge method